Fix My Requests
* Move useNewRequestPage hook to its own file
* Fix issue with missing first name in RequestActionDialog
